Fix shortestPath not getting min distance tile
The shortestPath function worked by getting the tile with the shortest
distance, and iterating on it, but since the minimum heap didn't update
when one of the tiles' distance was changed, a manual refresh was
needed. Now everytime distance is changed, the tile is removed and
added to the heap.